requirements.txt
streamlit
Pillow
fpdf2